Mirrors branch main from private at 9c7cceefc844b0fa24e163690bf064b7c54a8435

+++ src/gadgets/non_native_field/implementations/utils.rs
@@ -304,7 +304,7 @@ pub fn u1024_to_u16_words<const N: usize>(input: &U1024) -> [u16; N] {
     let mut result = [0u16; N];
     let mut tmp = *input;
     for dst in result.iter_mut() {
-        let low = tmp.as_words()[0] as u64;
+        let low = tmp.as_words()[0];
         *dst = low as u16;
 
         tmp = tmp.shr_vartime(16);
